pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "LaTeX CV rendering with Jinja2 and YAML"
requires-python = ">=3.11"
dependencies = [
  "Jinja2==3.1.4",
  "PyYAML==6.0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# No uv-specific config needed; using default groups


